+++ .config/rollup.config.js
@@ -18,6 +18,22 @@ export default [
         plugins: [commonjs(), resolve(), esbuild(), cleanup({ extensions: ['ts'] })],
         output: [
             {
+                file: './dist/index.js',
+                format: 'cjs',
+                sourcemap: true,
+            },
+            {
+                file: './dist/index.mjs',
+                format: 'es',
+                sourcemap: true,
+            },
+        ],
+    }),
+    bundle({
+        input: './src/lib/Db/index.ts',
+        plugins: [commonjs(), resolve(), esbuild(), cleanup({ extensions: ['ts'] })],
+        output: [
+            {
                 file: './dist/db/index.js',
                 format: 'cjs',
                 sourcemap: true,
@@ -30,11 +46,11 @@ export default [
         ],
     }),
     bundle({
-        input: './src/lib/evaluators/index.ts',
+        input: './src/lib/Evaluators/index.ts',
         plugins: [commonjs(), resolve(), esbuild(), cleanup({ extensions: ['ts'] })],
         output: [
             {
-                file: './dist/evaluators/index.js',
+                file: './dist/valuators/index.js',
                 format: 'cjs',
                 sourcemap: true,
             },
@@ -48,13 +64,21 @@ export default [
     bundle({
         input: './src/index.ts',
         output: {
+            file: './dist/types/index.d.ts',
+            format: 'es',
+        },
+        plugins: [resolve(), commonjs(), cleanup({ extensions: ['.ts'] }), dts()],
+    }),
+    bundle({
+        input: './src/lib/Db/index.ts',
+        output: {
             file: './dist/types/db.d.ts',
             format: 'es',
         },
         plugins: [resolve(), commonjs(), cleanup({ extensions: ['.ts'] }), dts()],
     }),
     bundle({
-        input: './src/lib/evaluators/index.ts',
+        input: './src/lib/Evaluators/index.ts',
         output: {
             file: './dist/types/evaluators.d.ts',
             format: 'es',
