+++ src/index.ts
@@ -1,6 +1,19 @@
 /** @format */
 
-export { Flotsam } from './lib';
+export {
+    Flotsam,
+    Exactly,
+    GreaterThan,
+    GreaterThanOrEqual,
+    In,
+    Is,
+    LessThan,
+    LessThanOrEqual,
+    Like,
+    NotEqual,
+    RegExp,
+    Unsafe,
+} from './lib';
 export type { Collection, ObjectId, JSONDocument } from './lib';
 export type {
     FlotsamInit,
@@ -12,4 +25,8 @@ export type {
     DocumentInit,
     QueueTask,
     Rejector,
+    ErrorHandler,
+    EvaluatorOptions,
+    FindByProperty,
+    FindOptions,
 } from './types';
