+++ src/lib/Db/evaluateFindOptions.ts
@@ -1,8 +1,7 @@
 /** @format */
 
 import { type } from 'os';
-import { FindOptions, Document } from '../types';
-import { FindByProperty } from '../types/FindByProperty';
+import { FindOptions, Document, FindByProperty } from '../../types';
 
 /**
  * @description
